requirements/bot.txt
aiogram
aiocache
uvloop

requests
fake-useragent
beautifulsoup4
lxml

llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-program-openai

pytest
